directory-refactor: adding .py files for jobs
